use factory from titiler.xarray
