+++ src/titiler/xarray_api/__init__.py
@@ -1,0 +1,3 @@
+"""titiler.xarray_api"""
+
+__version__ = "0.3.0"
