+++ src/titiler/xarray_api/map.html
