+++ src/titiler/xarray_api/middleware.py
