+++ src/titiler/xarray_api/redis_pool.py
@@ -1,10 +1,11 @@
-""" Redis singleton class. """
+"""Redis singleton class."""
+
 import os
 
 import fakeredis
 import redis  # type: ignore
 
-from titiler.xarray.settings import ApiSettings
+from titiler.xarray_api.settings import ApiSettings
 
 api_settings = ApiSettings()
 
@@ -24,7 +25,7 @@ class RedisCache:
         return cls._instance
 
 
-def get_redis():
+def get_redis() -> redis.Redis:
     """Get a redis connection."""
     if os.getenv("TEST_ENVIRONMENT"):
         server = fakeredis.FakeServer()
