+++ src/titiler/xarray_api/settings.py
