+++ tests/conftest.py
@@ -10,7 +10,7 @@ def app(monkeypatch):
     monkeypatch.setenv("TITILER_XARRAY_DEBUG", "TRUE")
     monkeypatch.setenv("TEST_ENVIRONMENT", "1")
 
-    from titiler.xarray.main import app
+    from titiler.xarray_api.main import app
 
     with TestClient(app) as client:
         yield client
