+++ tests/fixtures/responses/pyramid_zarr_info.json
@@ -5,10 +5,19 @@
     180.72289156626505,
     90.36144578313252
   ],
-  "minzoom": 0,
-  "maxzoom": 0,
-  "band_metadata": [],
-  "band_descriptions": [],
+  "crs": "http://www.opengis.net/def/crs/EPSG/0/4326",
+  "band_metadata": [
+    [
+      "b1",
+      {}
+    ]
+  ],
+  "band_descriptions": [
+    [
+      "b1",
+      "value"
+    ]
+  ],
   "dtype": "float64",
   "nodata_type": "Nodata",
   "name": "value",
