+++ tests/fixtures/responses/test_zarr_store_info.json
@@ -1,7 +1,5 @@
 {
     "bounds": [-182.5, -92.57142857142857, 177.5, 92.57142857142857],
-    "minzoom": 0,
-    "maxzoom": 0,
     "band_metadata": [],
     "band_descriptions": [],
     "dtype": "float64",
