+++ tests/fixtures/responses/test_zarr_store_zarr_info.json
@@ -5,10 +5,19 @@
     177.5,
     92.57142857142857
   ],
-  "minzoom": 0,
-  "maxzoom": 0,
-  "band_metadata": [],
-  "band_descriptions": [],
+  "crs": "http://www.opengis.net/def/crs/EPSG/0/4326",
+  "band_metadata": [
+    [
+      "b1",
+      {}
+    ]
+  ],
+  "band_descriptions": [
+    [
+      "b1",
+      "value"
+    ]
+  ],
   "dtype": "uint8",
   "nodata_type": "None",
   "name": "CDD0",
