+++ tests/test_app.py
@@ -5,7 +5,6 @@ from helpers import find_string_in_stream
 
 DATA_DIR = "tests/fixtures"
 test_zarr_store = os.path.join(DATA_DIR, "test_zarr_store.zarr")
-test_reference_store = os.path.join(DATA_DIR, "reference.json")
 test_netcdf_store = os.path.join(DATA_DIR, "testfile.nc")
 test_unconsolidated_store = os.path.join(DATA_DIR, "unconsolidated.zarr")
 test_pyramid_store = os.path.join(DATA_DIR, "pyramid.zarr")
@@ -15,15 +14,6 @@ test_zarr_store_params = {
     "variables": ["CDD0", "DISPH", "FROST_DAYS", "GWETPROF"],
 }
 
-test_reference_store_params = {
-    "params": {
-        "url": test_reference_store,
-        "variable": "value",
-        "reference": True,
-        "decode_times": False,
-    },
-    "variables": ["value"],
-}
 test_netcdf_store_params = {
     "params": {"url": test_netcdf_store, "variable": "data", "decode_times": False},
     "variables": ["data"],
@@ -65,10 +55,6 @@ def test_get_variables_test(app):
     return get_variables_test(app, test_zarr_store_params)
 
 
-def test_get_variables_reference(app):
-    return get_variables_test(app, test_reference_store_params)
-
-
 def test_get_variables_netcdf(app):
     return get_variables_test(app, test_netcdf_store_params)
 
@@ -87,8 +73,10 @@ def get_info_test(app, ds_params):
         params=ds_params["params"],
     )
     assert response.status_code == 200
+    print(json.dumps(response.json(), indent=2))
+    expectation_fn = f"{ds_params['params']['url'].replace(DATA_DIR, f'{DATA_DIR}/responses').replace('.', '_')}_info.json"
     with open(
-        f"{ds_params['params']['url'].replace(DATA_DIR, f'{DATA_DIR}/responses').replace('.', '_')}_info.json",
+        expectation_fn,
         "r",
     ) as f:
         assert response.json() == json.load(f)
@@ -96,10 +84,6 @@ def get_info_test(app, ds_params):
 
 def test_get_info_test(app):
     return get_info_test(app, test_zarr_store_params)
-
-
-def test_get_info_reference(app):
-    return get_info_test(app, test_reference_store_params)
 
 
 def test_get_info_netcdf(app):
@@ -116,12 +100,15 @@ def test_get_info_pyramid(app):
 
 def get_tilejson_test(app, ds_params):
     response = app.get(
-        "/tilejson.json",
+        "/WebMercatorQuad/tilejson.json",
         params=ds_params["params"],
     )
     assert response.status_code == 200
+    print(json.dumps(response.json(), indent=2))
+    expectation_fn = f"{ds_params['params']['url'].replace(DATA_DIR, f'{DATA_DIR}/responses').replace('.', '_')}_tilejson.json"
+    print(expectation_fn)
     with open(
-        f"{ds_params['params']['url'].replace(DATA_DIR, f'{DATA_DIR}/responses').replace('.', '_')}_tilejson.json",
+        expectation_fn,
         "r",
     ) as f:
         assert response.json() == json.load(f)
@@ -129,10 +116,6 @@ def get_tilejson_test(app, ds_params):
 
 def test_get_tilejson_test(app):
     return get_tilejson_test(app, test_zarr_store_params)
-
-
-def test_get_tilejson_reference(app):
-    return get_tilejson_test(app, test_reference_store_params)
 
 
 def test_get_tilejson_netcdf(app):
@@ -149,7 +132,7 @@ def test_get_tilejson_pyramid(app):
 
 def get_tile_test(app, ds_params, zoom: int = 0):
     response = app.get(
-        f"/tiles/{zoom}/0/0.png",
+        f"/tiles/WebMercatorQuad/{zoom}/0/0.png",
         params=ds_params["params"],
     )
     assert response.status_code == 200
@@ -163,10 +146,6 @@ def get_tile_test(app, ds_params, zoom: int = 0):
 
 def test_get_tile_test(app):
     return get_tile_test(app, test_zarr_store_params)
-
-
-def test_get_tile_reference(app):
-    return get_tile_test(app, test_reference_store_params)
 
 
 def test_get_tile_netcdf(app):
@@ -185,7 +164,7 @@ def test_get_tile_pyramid(app):
 
 def test_get_tile_pyramid_error(app):
     response = app.get(
-        "/tiles/3/0/0.png",
+        "/tiles/WebMercatorQuad/3/0/0.png",
         params=test_pyramid_store_params["params"],
     )
     assert response.status_code == 422
@@ -207,10 +186,6 @@ def histogram_test(app, ds_params):
 
 def test_histogram_test(app):
     return histogram_test(app, test_zarr_store_params)
-
-
-def test_histogram_reference(app):
-    return histogram_test(app, test_reference_store_params)
 
 
 def test_histogram_netcdf(app):
@@ -238,7 +213,7 @@ def test_histogram_error(app):
                 "loc": ["query", "variable"],
                 "msg": "Field required",
                 "input": None,
-                "url": "https://errors.pydantic.dev/2.1.2/v/missing",
+                "url": "https://errors.pydantic.dev/2.10/v/missing",
             }
         ]
     }
