--- titiler/xarray/__init__.py
@@ -1,3 +1,0 @@
-"""titiler.xarray"""
-
-__version__ = "0.1.0"
